Getting and Cleaning Data Course Project
This project is presented for the purpose of submission in the Cousera Getting and Cleaning Data course. 
The purpose of this project is to demonstrate the student ability to collect, work with, and clean a data set. 
The goal is to prepare tidy data that can be used for later analysis.